fix: cqrs 패턴 적용 및 사장님 home view 기능 개발
