+++ src/main/java/com/jangburich/domain/order/domain/repository/CustomOrderRepository.java
@@ -1,0 +1,10 @@
+package com.jangburich.domain.order.domain.repository;
+
+import java.time.LocalDateTime;
+import java.util.List;
+
+import com.jangburich.domain.order.domain.Orders;
+
+public interface CustomOrderRepository {
+    List<Orders> queryOrdersByStoreIdAndStartDateAndEndDate(Long storeId, LocalDateTime startDate, LocalDateTime endDate);
+}
