+++ src/main/java/com/jangburich/domain/store/repository/StoreTeamRepository.java
@@ -6,10 +6,12 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
 
 import com.jangburich.domain.store.domain.Store;
 import com.jangburich.domain.store.domain.StoreTeam;
-import com.jangburich.domain.store.dto.response.store.StoreTeamResponseDTO;
+import com.jangburich.domain.store.presentation.dto.response.store.StoreTeamResponseDTO;
 
 public interface StoreTeamRepository extends JpaRepository<StoreTeam, Long> {
 	Optional<StoreTeam> findByStoreIdAndTeamId(Long store_id, Long team_id);
@@ -19,4 +21,9 @@ public interface StoreTeamRepository extends JpaRepository<StoreTeam, Long> {
     Optional<StoreTeam> findByStoreAndTeam(Store store, Team team);
 
 	List<StoreTeamResponseDTO> findAllByStoreOrderByCreatedAtDesc(Store store);
+
+	@Query("SELECT st FROM StoreTeam st " +
+		"JOIN FETCH st.store s " +
+		"WHERE s.id = :storeId")
+	List<StoreTeam> findByStoreIdWithStoreAndTeam(@Param("storeId") Long storeId);
 }
